Don’t modify inputs in-place! Fixes #10.
This changes topojson.quantize to make a shallow copy of the input topology,
such that the input topology is no longer modified in-place. (Currently this
loses the id, bbox and properties of objects; I will fix this and add tests in a
subsequent commit.)

This also changes topojson.bbox so that it always computes and returns a new
bounding box for the given topology, rather than assigning the topology.bbox
property if missing.

Likewise, the transform functions returned by topojson.[un]transform now return
a copy of the input point rather than modifying it. These methods have also been
changed to take a transform object as input, rather than a topology object. The
topojson.quantize method now uses topojson.untransform internally.